learn-go-with-tests/06-pointers-and-errors/wallet_test.go
package wallet

import (
	"testing"
)

func TestWallet(t *testing.T) {
	assertBal := func(t testing.TB, wallet Wallet, want Bitcoin) {
		t.Helper()
		got := wallet.Balance()
		if got != want {
			t.Errorf("got %s, want %s", got, want)
		}
	}
	assertErr := func(t testing.TB, err error) {
		t.Helper()
		if err == nil {
			t.Error("wanted an error but didn't get one.")
		}
	}
	t.Run("deposit", func(t *testing.T) {
		wallet := Wallet{}
		wallet.Deposit(Bitcoin(10))
		assertBal(t, wallet, Bitcoin(10))
	})
	t.Run("withdraw", func(t *testing.T) {
		wallet := Wallet{balance: Bitcoin(20)}
		wallet.Withdraw(Bitcoin(10))
		assertBal(t, wallet, Bitcoin(10))
	})
	t.Run("withdraw insufficient funds", func(t *testing.T) {
		initialBal := Bitcoin(20)
		wallet := Wallet{initialBal}
		err := wallet.Withdraw(Bitcoin(30))
		assertErr(t, err)
		assertBal(t, wallet, initialBal)
	})
}
